fix: [FE] 카카오톡 공유하기 결과화면 전송
